Fixed bug causing Dublin Core metadata fields to be listed as empty.
Added a list of ancestor objects to SOLR during an object save.

Split the object's descMetadata from it's full metadata snippet so that the metadata of it's children. This is needed as OM proxies can only produce relative paths that start with "//" causing the object's children's metadata to get listed as the object's metadata, e.g the "//" will pick up the children's title field as well as object's title field and merge the fields into one result. Also, the full XML file can be quite large, e.g. NIVAL is about 11 MB in size!, so removing the bulk of it from the OM terminologies should speed things up. The full snippet is now stored in a datastream called fullMetadata, with no indexing of it's content.

Added an is_preceded_by predicate for relating ordered siblings together.

Replaced calls to new? with new_record? for all objects that inherit ActiveFedora::Base.

Started finalizing unit tests for EAD.